fix!: User warning
BREAKING-CHANGE: Complete restructure of user warnings, all user warnings data will be dropped.

+++ src/main/kotlin/be/duncanc/discordmodbot/bot/sequences/RevokeWarnPoints.kt
@@ -1,7 +1,7 @@
 package be.duncanc.discordmodbot.bot.sequences
 
 import be.duncanc.discordmodbot.bot.commands.CommandModule
-import be.duncanc.discordmodbot.data.entities.GuildWarnPoints
+import be.duncanc.discordmodbot.data.entities.GuildWarnPoint
 import be.duncanc.discordmodbot.data.repositories.jpa.GuildWarnPointsRepository
 import net.dv8tion.jda.api.Permission
 import net.dv8tion.jda.api.entities.Message
@@ -30,32 +30,31 @@ class RevokeWarnPoints(
             throw IllegalArgumentException("A mention is required to use this command", stringIndexOutOfBoundsException)
         }
         val userGuildWarnPoints =
-            guildWarnPointsRepository.findById(GuildWarnPoints.GuildWarnPointsId(userId, event.guild.idLong))
-        userGuildWarnPoints.ifPresentOrElse(
-            {
-                event.jda.addEventListener(RevokePointsSequence(event.author, event.channel, it))
-            },
-            {
-                event.channel.sendMessage("This user id has no warnings or the user does not exist.").queue {
-                    it.delete().queueAfter(1, TimeUnit.MINUTES)
-                }
+            guildWarnPointsRepository.findAllByGuildIdAndUserId(userId, event.guild.idLong)
+
+        if (userGuildWarnPoints.isNotEmpty()) {
+            event.jda.addEventListener(RevokePointsSequence(event.author, event.channel, userGuildWarnPoints))
+        } else {
+            event.channel.sendMessage("This user id has no warnings or the user does not exist.").queue {
+                it.delete().queueAfter(1, TimeUnit.MINUTES)
             }
-        )
+        }
     }
 
     inner class RevokePointsSequence(
         user: User,
         channel: MessageChannel,
-        private val userGuildWarnPoints: GuildWarnPoints
+        private val guildWarnPoints: Collection<GuildWarnPoint>,
     ) : Sequence(
         user,
         channel
     ), MessageSequence {
         init {
             val messageBuilder = StringBuilder("The user has had the following warnings in the past:\n\n")
-            userGuildWarnPoints.points.forEach { userWarnPoints ->
-                messageBuilder.append("$userWarnPoints\n\n")
-            }
+            guildWarnPoints
+                .forEach { userWarnPoints ->
+                    messageBuilder.append("Id ${userWarnPoints.id} - Reason: ${userWarnPoints.reason} - Created by <@${userWarnPoints.creatorId}>\n\n")
+                }
             messageBuilder.append("\n\nPlease enter the ID of the warning to revoke it.")
 
             SplitUtil.split(messageBuilder.toString(), Message.MAX_CONTENT_LENGTH, SplitUtil.Strategy.NEWLINE)
@@ -66,20 +65,13 @@ class RevokeWarnPoints(
 
         override fun onMessageReceivedDuringSequence(event: MessageReceivedEvent) {
             val selectedUUID: UUID = UUID.fromString(event.message.contentRaw)
-            val selectedUserWarnPoints = userGuildWarnPoints.points.find { it.id == selectedUUID }
-            if (selectedUserWarnPoints != null) {
-                userGuildWarnPoints.points.remove(selectedUserWarnPoints)
-                guildWarnPointsRepository.save(userGuildWarnPoints)
-                channel.sendMessage("The warning has been revoked.").queue {
-                    it.delete().queueAfter(1, TimeUnit.MINUTES)
-                }
-                super.destroy()
-            } else {
-                channel.sendMessage("The warning with that id was not found.").queue {
-                    it.delete().queueAfter(1, TimeUnit.MINUTES)
-                }
-                super.destroy()
+
+            guildWarnPointsRepository.deleteAllById(selectedUUID)
+
+            channel.sendMessage("The warning has been revoked.").queue {
+                it.delete().queueAfter(1, TimeUnit.MINUTES)
             }
+            super.destroy()
         }
     }
 }
